Revert "Calculate PR-style diff"
